requirements.txt
openai>=1.0.0
SpeechRecognition>=3.10.0
pynput>=1.7.6
pyaudio>=0.2.11
python-dotenv>=1.0.0

